ENH: Build Slicer module without GPL code from "gradientWidgetQT"
This commit replaces "ShapePopulationGradientWidgetQT.h" header now
referenced in the ShapePopulationQT.ui file.
* If the option ShapePopulationViewer_BUILD_SLICER_EXTENSION is ON,
  it derives from "ctkVTKScalarsToColorsWidget".
* Otherwise it derives from "gradientWidgetQT".

When building against Slicer, the use of "ctkVTKScalarsToColorsWidget"
also greatly simplifies the management of color bar points, there is no
need to normalize the values.

Once the standalone CLI is also updated to depend on CTK, setting the scalar
bar should be a matter of associating the "vtkColorTransferFunction" from
the widget directly with the "vtkScalarBarActor".